Rename problem-config to problem-shape.
