+++ src/SConscript
@@ -76,7 +76,7 @@ model/buffer.cpp
 model/topology.cpp
 util/numeric.cpp
 util/map2d.cpp
-workload/problem-config.cpp
+workload/problem-shape.cpp
 workload/workload-config.cpp
 workload/operation-space.cpp
 """)
@@ -96,7 +96,7 @@ model/buffer.cpp
 model/topology.cpp
 util/numeric.cpp
 util/map2d.cpp
-workload/problem-config.cpp
+workload/problem-shape.cpp
 workload/workload-config.cpp
 workload/operation-space.cpp
 """)
@@ -113,7 +113,7 @@ model/buffer.cpp
 model/topology.cpp
 util/numeric.cpp
 util/map2d.cpp
-workload/problem-config.cpp
+workload/problem-shape.cpp
 workload/workload-config.cpp
 workload/operation-space.cpp
 """)
