+++ src/loop-analysis/tiling.hpp
@@ -31,7 +31,7 @@
 
 #include "mapping/loop.hpp"
 #include "util/numeric.hpp"
-#include "workload/problem-config.hpp"
+#include "workload/problem-shape.hpp"
 
 namespace tiling
 {
