+++ src/mapspaces/mapspace-factory.hpp
@@ -30,10 +30,7 @@
 #include <libconfig.h++>
 
 #include "mapspaces/uber.hpp"
-//#include "mapspaces/eyeriss.hpp"
-//#include "mapspaces/craft.hpp"
-//#include "mapspaces/bollywood.hpp"
-#include "workload/problem-config.hpp"
+#include "workload/problem-shape.hpp"
 
 namespace mapspace
 {
@@ -55,20 +52,6 @@ MapSpace* ParseAndConstruct(libconfig::Setting& config,
   {
     mapspace = new Uber(config, arch_specs, workload_config);
   }
-  // else if (mapspace_template == "eyeriss")
-  // {
-  //   assert(false);
-  //   //mapspace = new Eyeriss(config, workload_config);
-  // }
-  // else if (mapspace_template == "CRAFT")
-  // {
-  //   mapspace = new CRAFT(config, arch_specs, workload_config);
-  // }
-  // else if (mapspace_template == "bollywood")
-  // {
-  //   assert(false);
-  //   //mapspace = new Bollywood(config, workload_config);
-  // }
   else
   {
     std::cerr << "ERROR: unsupported mapspace template: " << mapspace_template << std::endl;
